change folder name and add validations
