+++ src/main/java/org/_p1m/foodorderingsystem/features/restaurant/dto/response/RestaurantUpdateResponseDto.java
@@ -1,14 +1,12 @@
-package org._p1m.foodorderingsystem.features.restaurant_registration.dto.response;
+package org._p1m.foodorderingsystem.features.restaurant.dto.response;
 
 import lombok.Data;
 
 @Data
-public class RestaurantResponseDto {
-
+public class RestaurantUpdateResponseDto {
     private String restaurantName;
     private String contactNumber;
     private String restaurantImage;
     private String nrc;
     private String kpayNumber;
-    private Long resOwnerId;
 }
