--- src/main/java/org/_p1m/foodorderingsystem/features/restaurant_registration/dto/response/RestaurantUpdateResponseDto.java
@@ -1,12 +1,0 @@
-package org._p1m.foodorderingsystem.features.restaurant_registration.dto.response;
-
-import lombok.Data;
-
-@Data
-public class RestaurantUpdateResponseDto {
-    private String restaurantName;
-    private String contactNumber;
-    private String restaurantImage;
-    private String nrc;
-    private String kpayNumber;
-}
